backend/requirements.txt
# Backend Python dependencies for GMB Calls Analyzer
# Install all at once: pip install -r requirements.txt

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Google Gemini API client (for Insights generation)
google-genai>=1.0.0

# Environment variables (.env file loader)
python-dotenv>=1.0.0

# ── AI Trainer subsystem (gated behind TRAINER_ENABLED flag) ─────────────────
# Not used unless the trainer is enabled at boot. See AITrainer_TechPlan_v1.md A1.
pandas>=2.2
portalocker>=2.10
python-multipart>=0.0.9
websockets>=13

# Test runners (used only by `pytest backend/tests/`)
pytest>=8
pytest-asyncio>=0.24
httpx>=0.27

# Note: csv, json, os, re, pathlib, typing are part of the Python standard library — no install needed.
